requirements.txt
Flask
Werkzeug
python-dotenv 